Required pom file
