--- f2.txt
@@ -1,1 +1,0 @@
-Hello there Good Evening!
