+++ src/main/java/com/qa/runtimeterror2/Monster.java
@@ -1,9 +1,8 @@
-package runtimeterror2;
+package com.qa.runtimeterror2;
 
-public class GameEntity {
+public class Monster extends GameEntity{
 	
-	// location
-	// unique message
+	public String uniqueMessage;
 	public static int posX;
 	public static int posY;
 	
@@ -18,13 +17,18 @@ public class GameEntity {
 	}
 	
 	// TO BE IMPLEMENTED
-	public void setPosition(int pos[]) {
+	public  void setPosition(int pos[]) {
 		// description: player posX=0, posY=0
 		posX=pos[0];
 		posY=pos[1];
 	}
 		
 	
-
+	
+	// TO BE IMPLEMENTED
+	public void displayUniqueMessage() {
+		// description: output message
+		System.out.println("You've hit a monster! Go back to start.");
+	}
 
 }
