+++ src/main/java/com/qa/runtimeterror2/UserInput.java
@@ -1,4 +1,4 @@
-package runtimeterror2;
+package com.qa.runtimeterror2;
 
 import java.util.Scanner;
